First commit
Task 1:
Create classes "Director", "Worker", "Accountant".
Implement an interface with a method that specifies the behavior for displaying the job title.
Implement this method in the created classes.
Create a runner class with a main method to run the program.
Request the job code from the console and display the job title on the screen.

Task 2:
Write the hierarchy of the classes Shape, Triangle, Rectangle, Circle.
Implement a function for calculating the area for each type of figure and calculating the perimeter (using abstract class/methods).
Create an array of 5 different shapes.
Loop through the array and display information about each shape.
Display the sum of the perimeters of all shapes in the array.